+++ osu.Game/Overlays/Chat/ChatLine.cs
@@ -200,7 +200,15 @@ namespace osu.Game.Overlays.Chat
             timestamp.FadeTo(message is LocalEchoMessage ? 0 : 1, 500, Easing.OutQuint);
 
             timestamp.Text = $@"{message.Timestamp.LocalDateTime:HH:mm:ss}";
-            username.Text = $@"{message.Sender.Username}" + (senderHasBackground || message.IsAction ? "" : ":");
+
+            if ( message.IsAction )
+            {
+                username.Text = @"* " + $"{message.Sender.Username}" + ":";
+            }
+            else
+            {
+                username.Text = $@"{message.Sender.Username}" + ":";
+            }
 
             // remove non-existent channels from the link list
             message.Links.RemoveAll(link => link.Action == LinkAction.OpenChannel && chatManager?.AvailableChannels.Any(c => c.Name == link.Argument) != true);