+++ osu.Game/Overlays/Toolbar/Toolbar.cs
@@ -70,7 +70,7 @@ namespace osu.Game.Overlays.Toolbar
                     AutoSizeAxes = Axes.X,
                     Children = new Drawable[]
                     {
-                        //new ToolbarVideoButton(),
+                        new ToolbarTimeButton(),
                         new ToolbarChangelogButton(),
                         new ToolbarDirectButton(),
                         new ToolbarChatButton(),