+++ osu.Game/Screens/Menu/ButtonSystem.cs
@@ -126,6 +126,7 @@ namespace osu.Game.Screens.Menu
             buttonsPlay.Add(new Button(@"单人游戏", @"button-solo-select", FontAwesome.Solid.User, new Color4(102, 68, 204, 255), () => OnSolo?.Invoke(), WEDGE_WIDTH, Key.P));
             buttonsPlay.Add(new Button(@"多人游戏", @"button-generic-select", FontAwesome.Solid.Users, new Color4(94, 63, 186, 255), onMulti, 0, Key.M));
             buttonsPlay.Add(new Button(@"表格", @"button-generic-select", OsuIcon.Charts, new Color4(80, 53, 160, 255), () => OnChart?.Invoke()));
+            buttonsPlay.Add(new Button(@"关于该分支", @"button-generic-select", FontAwesome.Solid.Boxes, new Color4(80, 53, 160, 255), OnMf ));
             buttonsPlay.ForEach(b => b.VisibleState = ButtonSystemState.Play);
 
             buttonsTopLevel.Add(new Button(@"游玩", @"button-play-select", OsuIcon.Logo, new Color4(102, 68, 204, 255), () => State = ButtonSystemState.Play, WEDGE_WIDTH, Key.P));
@@ -152,6 +153,15 @@ namespace osu.Game.Screens.Menu
             if (idleTracker != null) isIdle.BindTo(idleTracker.IsIdle);
 
             sampleBack = audio.Samples.Get(@"Menu/button-back-select");
+        }
+
+        private void OnMf(){
+            notifications?.Post(new SimpleNotification
+                {
+                    Text =  "暂未实现qwq",
+                    Icon = FontAwesome.Solid.Boxes,
+                });
+            return;
         }
 
         private void onMulti()